The unittest will now send random amounts of data with the packet header to OAServer.  I've been able to test out receiving the data, call the process function, dynamically allocating more space for big packets, and using multiple threads.  Going to test out the checksum valid checking

+++ include/Util.h
@@ -26,7 +26,7 @@ class Util {
    ~Util(void);
 
     /* Make this a template so any data type can be used */
-    unsigned short ComputeChecksum(unsigned short *pBuff, int mNumBytes);
+    int ComputeChecksum(int *pBuff, int mNumBytes);
 
 #if 0
     template <class A>
